Move round name mapping into constant file.
